+++ Minceraft 2.0/src/Source Files/Game.cpp
@@ -15,7 +15,7 @@ namespace Game {
 				break;
 			auto end_time = std::chrono::high_resolution_clock::now();
 			auto duration = std::chrono::duration_cast<std::chrono::microseconds>(end_time - start_time);
-			delta_time = (float)((duration.count()) / 1000000.0);
+			delta_time = (double)((duration.count()) / 1000000.0);
 			// calculate FPS
 			static int count = 0;
 			static float total_time = 0;
